frontend: Extract prepareRoutes from Sidebar
